Category Item improved. Add product improved

+++ app/src/main/java/com/symbol/shoppinglist/ui/categoriesManage/ManageCategories.kt
@@ -1,26 +1,29 @@
 package com.symbol.shoppinglist.ui.categoriesManage
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Card
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Delete
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.symbol.shoppinglist.IconName
 import com.symbol.shoppinglist.database.local.entities.Category
 import com.symbol.shoppinglist.navigation.CategoriesDirections
-import com.symbol.shoppinglist.ui.ColorSquare
+import com.symbol.shoppinglist.ui.theme.Shapes
 
 @Composable
 fun ManageCategories(
@@ -43,23 +46,46 @@ fun ManageCategories(
 fun CategoryItem(
     modifier: Modifier = Modifier,
     category: Category,
-    onClick: (Int) -> Unit,
-    deleteIconClick: (Category) -> Unit
+    onClick: ((Int) -> Unit)? = null,
+    deleteIconClick: ((Category) -> Unit)? = null,
+    content: @Composable (() -> Unit)? = null
 ) {
-    Row(modifier = modifier.clickable {
-        onClick(category.id)
-    }
+    val chooseModifier =
+        if (onClick != null) Modifier.clickable { onClick(category.id) } else Modifier
+    Card(
+        modifier = chooseModifier
+            .fillMaxWidth()
+            .padding(horizontal = 4.dp, vertical = 2.dp),
+        elevation = 3.dp,
+        backgroundColor = Color(category.color)
     ) {
-        Text(modifier = modifier, text = category.name)
-        Box(
-            modifier = modifier
-                .clip(RoundedCornerShape(20))
+        Row(
+            modifier = Modifier
+                .fillMaxWidth(0.9f)
+                .padding(4.dp),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
         ) {
-            ColorSquare(modifier, category.color)
+            Text(
+                modifier = Modifier
+                    .align(Alignment.CenterVertically)
+                    .padding(horizontal = 4.dp),
+                text = category.name,
+                textAlign = TextAlign.Left
+            )
+            if (deleteIconClick != null) {
+                Icon(
+                    Icons.Rounded.Delete, IconName.DELETE,
+                    Modifier
+                        .clickable {
+                            deleteIconClick(category)
+                        }
+                        .align(Alignment.CenterVertically)
+                )
+            }
+            if(content != null)
+            content()
         }
-        Icon(Icons.Rounded.Delete, IconName.DELETE, modifier.clickable {
-            deleteIconClick(category)
-        })
     }
 }
 
@@ -70,15 +96,21 @@ fun ListOfCategories(
     onClick: (Int) -> Unit,
     deleteIconClick: (Category) -> Unit
 ) {
-    LazyColumn(
-        modifier = modifier
-    ) {
-        items(categories, key = { it.id }) { item ->
-            CategoryItem(
-                category = item,
-                onClick = onClick,
-                deleteIconClick = deleteIconClick
-            )
+    Card(modifier = Modifier
+        .fillMaxWidth()
+        .padding(8.dp),
+    elevation = 10.dp) {
+        LazyColumn(
+            modifier = modifier
+        ) {
+            items(categories, key = { it.id }) { item ->
+                CategoryItem(
+                    category = item,
+                    onClick = onClick,
+                    deleteIconClick = deleteIconClick,
+                )
+                Spacer(modifier = Modifier.height(3.dp))
+            }
         }
     }
 }
